Converted files to TS, added tsconfig and other settings
