+++ packages/timelapse-visual/package.json
@@ -22,20 +22,27 @@
   "dependencies": {
     "@splunk/babel-preset": "3.0.0",
     "@splunk/react-ui": "^4.7.0",
+    "@splunk/search-job": "^2.0.0",
     "@splunk/themes": "^0.8.0",
     "react-datetime": "^3.1.1",
     "webpack-cli": "3.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.2.0",
+    "@marenaud/react-timeline-range-slider": "1.4.1",
     "@splunk/eslint-config": "^4.0.0",
     "@splunk/react-icons": "3.2.0",
     "@splunk/splunk-utils": "^2.1.0",
+    "@splunk/ui-utils": "^1.4.0",
     "@splunk/webpack-configs": "^5.0.0",
-    "@marenaud/react-timeline-range-slider": "1.4.1",
+    "@types/jest": "^28.1.1",
+    "@types/node": "^17.0.41",
+    "@types/react": "^18.0.12",
+    "@types/react-dom": "^18.0.5",
     "babel-eslint": "^10.1.0",
     "babel-loader": "^8.0.4",
     "chai": "^3.5.0",
+    "date-fns": "2.24.0",
     "enzyme": "^3.11.0",
     "enzyme-adapter-react-16": "^1.15.2",
     "eslint": "^7.14.0",
@@ -43,15 +50,18 @@
     "eslint-config-prettier": "^6.15.0",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.21.5",
+    "eslint-plugin-react": "7.21.5",
     "eslint-plugin-react-hooks": "^4.2.0",
     "html-webpack-plugin": "^3.2.0",
     "jest": "^25.1.0",
-    "date-fns": "2.24.0",
+    "prettier": "^2.6.2",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
     "styled-components": "^5.1.1",
     "stylelint": "^13.0.0",
+    "ts-loader": "^9.3.0",
+    "tsdx": "^0.14.1",
+    "typescript": "^4.7.3",
     "webpack": "^4.16.2",
     "webpack-dev-server": "^3.1.5",
     "webpack-merge": "^4.1.3"
@@ -61,6 +71,6 @@
     "styled-components": "^5.1"
   },
   "engines": {
-    "node": ">=6"
+    "node": ">=14"
   }
 }
