+++ packages/timelapse-visual/src/TimelapseControls/timecontext.ts
