+++ packages/timelapse-visual/src/types/splunkTypes.ts
@@ -1,0 +1,3 @@
+export type QS = {
+  [key: string]: string | number | boolean;
+};
