+++ packages/timelapse-visual/tests/Documentation.unit.jsx
@@ -1,30 +1,28 @@
-/* eslint-env jest */
-
-import React from 'react';
+import React from "react";
 import { assert } from 'chai';
 import Enzyme, { mount } from 'enzyme';
 import EnzymeAdapterReact16 from 'enzyme-adapter-react-16';
-import DashboardSelector from '../DashboardSelector';
+import Documentation from '../src/Documentation/Documentation';
 
 // This sets up the enzyme adapter
 const adapter = new EnzymeAdapterReact16();
 Enzyme.configure({ adapter });
 
-describe('DashboardSelector', () => {
+describe('Documentation', () => {
     it('renders with default name', () => {
-        const wrapper = mount(<DashboardSelector />);
+        const wrapper = mount(<Documentation />);
         assert.include(wrapper.text(), 'Hello, User!');
         wrapper.unmount();
     });
 
     it('renders with custom name', () => {
-        const wrapper = mount(<DashboardSelector name="World" />);
+        const wrapper = mount(<Documentation name="World" />);
         assert.include(wrapper.text(), 'Hello, World!');
         wrapper.unmount();
     });
 
     it('increases the counter when button is clicked', () => {
-        const wrapper = mount(<DashboardSelector name="World" />);
+        const wrapper = mount(<Documentation name="World" />);
         assert.equal(wrapper.state('counter'), 0);
         wrapper.find('button').simulate('click');
         assert.equal(wrapper.state('counter'), 1);
