+++ packages/timelapse-visual/tests/ListDashboard.unit.jsx
@@ -1,30 +1,29 @@
 /* eslint-env jest */
-
-import React from 'react';
+import React from "react";
 import { assert } from 'chai';
 import Enzyme, { mount } from 'enzyme';
 import EnzymeAdapterReact16 from 'enzyme-adapter-react-16';
-import Listdashboards from '../Listdashboards';
+import ListDashboard from '../src/ListDashboard/ListDashboard';
 
 // This sets up the enzyme adapter
 const adapter = new EnzymeAdapterReact16();
 Enzyme.configure({ adapter });
 
-describe('Listdashboards', () => {
+describe('Listdashboard', () => {
     it('renders with default name', () => {
-        const wrapper = mount(<Listdashboards />);
+        const wrapper = mount(<ListDashboard />);
         assert.include(wrapper.text(), 'Hello, User!');
         wrapper.unmount();
     });
 
     it('renders with custom name', () => {
-        const wrapper = mount(<Listdashboards name="World" />);
+        const wrapper = mount(<ListDashboard name="World" />);
         assert.include(wrapper.text(), 'Hello, World!');
         wrapper.unmount();
     });
 
     it('increases the counter when button is clicked', () => {
-        const wrapper = mount(<Listdashboards name="World" />);
+        const wrapper = mount(<ListDashboard name="World" />);
         assert.equal(wrapper.state('counter'), 0);
         wrapper.find('button').simulate('click');
         assert.equal(wrapper.state('counter'), 1);
